pyproject.toml
[project]
name = "gemini-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Abdullah", email = "ma2404374@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.3.0",
    "google-generativeai>=0.8.4",
    "langchain-google-genai>=2.0.9",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]

[project.scripts]
gemini-practice = "gemini_practice:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
